Include MLJ as weakdep
