+++ Project.toml
@@ -7,7 +7,12 @@ version = "3.2.0-dev"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+
+[weakdeps]
 MLJ = "add582a8-e3ab-11e8-2d5e-e98b27df1bc7"
+
+[extensions]
+PreprocessMDMLJExt = "MLJ"
 
 [compat]
 DataFrames = "^1.3"
